test_app/users_office/serializers.py
from rest_framework import serializers
from users_office.models import *


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'


class UserRegistrSerializer(serializers.ModelSerializer):
    password2 = serializers.CharField()
    
    class Meta:
        model = User
        fields = ['email', 'password', 'password2']
 

    def save(self, *args, **kwargs):
        user = User(
            email=self.validated_data['email'],
        )
        password = self.validated_data['password']
        password2 = self.validated_data['password2']
        if password != password2:
            raise serializers.ValidationError({password: "Пароль не совпадает"})
        user.set_password(password)
        user.save()
        return user


class UserUpdateSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['first_name', 'last_name', 'phone', 'image']

    def update(self, instance, validated_data):
        instance.first_name = validated_data.get("first_name", instance.first_name)
        instance.last_name = validated_data.get("last_name", instance.last_name)
        instance.phone = validated_data.get("phone", instance.phone)
        instance.image = validated_data.get("image", instance.image)
        instance.save()
        return instance


class OfficeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Office
        fields = '__all__'
    
    def create(self, validated_data):
        return Office.objects.create(**validated_data)